requirements.txt
pandas
redis
rq
seqeval
beautifulsoup4
lxml
python-logstash
pymongo
pyyaml
simpletransformers
torch
